+++ src/App.vue
@@ -1,4 +1,7 @@
 <script setup lang="ts">
+import request from '~/composables/request'
+import { useUserStore } from '~/stores/user'
+
 // https://github.com/vueuse/head
 // you can use this to manipulate the document head in any components,
 // they will be rendered correctly in the html results with vite-ssg
@@ -7,6 +10,21 @@ useHead({
   meta: [
     { name: 'description', content: 'App for memorizing' },
   ],
+})
+onMounted(async() => {
+  const user = useUserStore()
+  const userData = localStorage.getItem('user')
+  if (userData) {
+    const { data, error } = await request.get('/user/refresh', true)
+    if (!error) {
+      user.refresh(data)
+      user.loggedIn = true
+      user.setData(JSON.parse(userData), false)
+    }
+  }
+  else {
+    user.refresh({ accessToken: '', refreshToken: '' })
+  }
 })
 </script>