+++ src/components/Header.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+<<<<<<< HEAD
 import { mapActions, storeToRefs } from 'pinia'
 import { toggleDark } from '~/composables'
 import { useLayoutStore } from '~/stores/layout'
@@ -7,6 +8,20 @@ const { t, availableLocales, locale } = useI18n()
 
 const { userPanelCollapsed } = storeToRefs(useLayoutStore())
 const { toggleUserPanel, toggleSidebar } = mapActions(useLayoutStore, ['toggleUserPanel', 'toggleSidebar'])
+=======
+import { storeToRefs } from 'pinia'
+// import { ref } from 'vue'
+import { toggleDark } from '~/composables'
+import { useLayoutStore } from '~/stores/layout'
+import { useUserStore } from '~/stores/user'
+const { t, availableLocales, locale } = useI18n()
+
+const layoutStore = useLayoutStore()
+const { userPanelCollapsed } = storeToRefs(layoutStore)
+const user = useUserStore()
+const { loggedIn: userLoggedIn, name: userName, picture: userPicture } = storeToRefs(user)
+// const { toggleUserPanel, toggleSidebar } = mapActions(useLayoutStore, ['toggleUserPanel', 'toggleSidebar'])
+>>>>>>> development
 
 const toggleLocales = () => {
   // change to some real logic
@@ -17,6 +32,7 @@ const toggleLocales = () => {
 
 <template>
   <header>
+<<<<<<< HEAD
     <button class="w-60px h-60px icon-btn !outline-none" :title="t('layout.toggle_sidebar')" @click="toggleSidebar">
       <div class="mx-auto" i="carbon-menu" />
     </button>
@@ -32,6 +48,40 @@ const toggleLocales = () => {
         <a class="icon-btn mx-2" :title="t('button.toggle_langs')" @click.stop.prevent="toggleLocales">
           <div i-carbon-language />
         </a>
+=======
+    <button class="w-60px h-60px icon-btn !outline-none" :title="t('layout.toggle_sidebar')" @click="layoutStore.toggleSidebar">
+      <div class="mx-auto" i="carbon-menu" />
+    </button>
+    <span id="logo"><router-link to="/">Reming</router-link></span>
+    <div id="userPanel" @click="layoutStore.toggleUserPanel">
+      <div v-if="userPanelCollapsed">
+        <div v-if="userLoggedIn" class="flex flex-row items-center gap-1em">
+          <router-link to="/profile" class="text-1.2em">
+            {{ userName }}
+          </router-link>
+          <img v-if="userPicture" id="userPicture" :src="`http://localhost:3001/${userPicture}`" alt="user picture">
+          <div v-else class="text-2em" i="carbon-user-avatar-filled-alt" />
+        </div>
+        <div v-else class="flex gap-5px">
+          <router-link to="/login" class="btn" @click.stop.prevent>
+            Log in
+          </router-link>
+          <router-link to="/signup" class="btn" @click.stop.prevent>
+            Sign up
+          </router-link>
+        </div>
+      </div>
+      <div v-else class="flex flex-row-reverse">
+        <button class="icon-btn mx-2" title="Log out" @click="user.logout">
+          <div i-carbon-logout />
+        </button>
+        <button class="icon-btn mx-2 !outline-none" :title="t('button.toggle_dark')" @click.stop.prevent="toggleDark()">
+          <div i="carbon-sun dark:carbon-moon" />
+        </button>
+        <button class="icon-btn mx-2" :title="t('button.toggle_langs')" @click.stop.prevent="toggleLocales">
+          <div i-carbon-language />
+        </button>
+>>>>>>> development
       </div>
     </div>
   </header>
@@ -88,4 +138,12 @@ const toggleLocales = () => {
     }
   }
 
+<<<<<<< HEAD
+=======
+  #userPicture {
+    height: 2em;
+    width: 2em;
+    border-radius: 50%;
+  }
+>>>>>>> development
 </style>