+++ src/components/Sidebar.vue
@@ -2,7 +2,12 @@
 import { storeToRefs } from 'pinia'
 import { useLayoutStore } from '~/stores/layout'
 
+<<<<<<< HEAD
 const { sidebarCollapsed } = storeToRefs(useLayoutStore())
+=======
+const layoutStore = useLayoutStore()
+const { sidebarCollapsed } = storeToRefs(layoutStore)
+>>>>>>> development
 
 // const toggleLocales = () => {
 //   // change to some real logic