+++ src/stores/layout.ts
@@ -1,5 +1,6 @@
 import { acceptHMRUpdate, defineStore } from 'pinia'
 
+<<<<<<< HEAD
 export const useLayoutStore = defineStore('layout', {
   state: () => ({
     sidebarCollapsed: true,
@@ -13,6 +14,39 @@ export const useLayoutStore = defineStore('layout', {
       this.userPanelCollapsed = !this.userPanelCollapsed
     },
   },
+=======
+interface PopupPayload {
+  type: 'error' | 'success' | 'info'
+  message: string
+  action?: {
+    text: string
+    action: Function
+  }
+}
+
+export const useLayoutStore = defineStore('layout', () => {
+  const sidebarCollapsed = ref(true)
+  const userPanelCollapsed = ref(true)
+
+  function toggleSidebar() {
+    sidebarCollapsed.value = !sidebarCollapsed.value
+  }
+  function toggleUserPanel() {
+    userPanelCollapsed.value = !userPanelCollapsed.value
+  }
+
+  function popup(payload: PopupPayload) {
+    alert(payload.message)
+  }
+
+  return {
+    sidebarCollapsed,
+    userPanelCollapsed,
+    toggleSidebar,
+    toggleUserPanel,
+    popup,
+  }
+>>>>>>> development
 })
 
 if (import.meta.hot)