+++ Changelog.md
@@ -1,6 +1,11 @@
 ## Reming changelog
 ---
+<<<<<<< HEAD
 **v0.0.0** - *05.05.22* 
+=======
+**v0.0** - *05.05.22*
+"Bootstrap"
+>>>>>>> development
 
 Client-side:
 * App created, based on [vitesse](https://github.com/antfu/vitesse) template
@@ -10,7 +15,21 @@ Server-side:
 * User authentication
 * Creating account, login, logout, getting profile, deleting account
 
+<<<<<<< HEAD
 To Do:
 * File uploading
 * User privacy settings
 * Querying users
+=======
+---
+**v0.1** - *09.05.22*
+"User"
+
+* Login, Signup, Pofile editing
+* Authentification, refreshing tokens, logout
+* File uploading
+* Basic responsive layout
+
+**0.2**
+"Materials"
+>>>>>>> development